requirements.txt
dlib==19.24.2
face_recognition==1.3.0
mysql_connector_repackaged==0.3.1
numpy==1.24.3
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.6.0.66
pandas==1.4.2
# Pillow==9.2.0
Pillow==10.0.0
PyYAML==5.4.1
# PyYAML==6.0.1
streamlit==1.24.0
tensorflow==2.9.1
tf_nightly==2.10.0.dev20220612
